requirements.txt
# Audio Event Detection - Dependencies
# Core ML
torch>=2.0.0
torchaudio>=2.0.0
timm>=0.9.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0

# Spectrograms / feature extraction
scipy>=1.10.0

# ML metrics
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0

# Data handling
pandas>=2.0.0

# Utilities
tqdm>=4.65.0
